+++ facebook-core/src/main/java/com/facebook/appevents/InternalAppEventsLogger.java
@@ -120,10 +120,6 @@ public class InternalAppEventsLogger {
         loggerImpl.flush();
     }
 
-    static String getSourceApplication() {
-        return AppEventsLoggerImpl.getSourceApplication();
-    }
-
     static Executor getAnalyticsExecutor() {
         return AppEventsLoggerImpl.getAnalyticsExecutor();
     }