+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Log Subscribe/StartTrial events going through GooglePlay store, which was logged as fb_mobile_purchase
 
+### Removed
+
+- Deprecate several activateApp and deactivateApp functions in AppEventsLogger.java
+
 ## [4.41.0] - 2019-03-08
 
 ### Removed